Move into yaml module
