+++ tests/test_example1.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from pydantic import BaseModel, Field
 
-from pydantic_examples import yaml_with_comments
+from pydantic_examples.yaml import yaml_with_comments
 
 
 class Nested(BaseModel):
